requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
Cython==3.0.11
filelock==3.16.1
finnhub-python==2.4.22
fsspec==2024.10.0
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
msgspec==0.18.6
nautilus_trader==1.208.0
networkx==3.4.2
numpy==2.2.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pandas==2.2.3
plotly==5.24.1
pyarrow==18.1.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
scikit-learn==1.6.0
scipy==1.14.1
setuptools==75.6.0
six==1.17.0
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
torch==2.5.1
tqdm==4.67.1
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvloop==0.21.0
pyyaml
